changes int to Long

+++ src/main/java/com/er/sagar/BlogAPI/entities/CommentsReply.java
@@ -12,7 +12,7 @@ public class CommentsReply {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     private String comment;
 
@@ -39,11 +39,11 @@ public class CommentsReply {
         this.comment = comment;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
